requirements.txt
PyQt6
pyxdg
google-generativeai
requests
cryptography
